Refactor/plugins introduction cleanup (#167)
This pull request introduces a major architectural migration plan for
integrating social evidence into the intelligence pipeline, alongside
improvements to the web UI code structure and stricter linting rules.
The most important changes are grouped below:

---

**Architecture & Migration Plan**

* Added a comprehensive migration plan
(`INTELLIGENCE_SOCIAL_CONVERGENCE_PLAN.md`) detailing the stepwise
integration of social evidence into intelligence, deprecation of the
standalone social domain, API transition, plugin architecture updates,
and compatibility strategies.

---

**Web UI Code Structure Improvements**

* Introduced a code assessment report highlighting misplaced logic in UI
components, hardcoded values, and potential dead code, with actionable
recommendations for refactoring and code cleanup.
* Extracted the `CollapsibleSection` UI component into its own file,
improving modularity and reusability in the codebase.
* Refactored the daily review candidates table into a dedicated
component (`DailyReviewCandidatesTable.tsx`), separating presentation
from business logic and supporting both mobile and desktop layouts.

---

**Linting & Code Quality**

* Increased strictness of React hook dependency linting by enabling
`'react-hooks/exhaustive-deps': 'error'` in the ESLint config, helping
prevent subtle bugs in effect dependencies.

---------
