+++ .gitignore
@@ -85,3 +85,4 @@ web-ui/.audit/
 # =========================
 test-results.xml
 coverage.xml
+docs/engineering/INTELLIGENCE_SOCIAL_CONVERGENCE_PLAN.md
