+++ web-ui/src/pages/Intelligence.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useMemo, useState } from 'react';
 import { RefreshCw } from 'lucide-react';
 import { useNavigate } from 'react-router-dom';
+import { useLocalStorage } from '@/hooks/useLocalStorage';
 import Card, { CardContent, CardHeader, CardTitle } from '@/components/common/Card';
 import Button from '@/components/common/Button';
 import IntelligenceOpportunityCard from '@/components/domain/intelligence/IntelligenceOpportunityCard';
+import IntelligenceConfigPanel from '@/components/domain/intelligence/IntelligenceConfigPanel';
 import {
-  useCreateIntelligenceSymbolSetMutation,
-  useDeleteIntelligenceSymbolSetMutation,
   useIntelligenceConfigQuery,
   useIntelligenceEventsQuery,
   useIntelligenceMetricsQuery,
@@ -14,69 +14,52 @@ import {
   useIntelligenceProvidersQuery,
   useIntelligenceRunStatus,
   useIntelligenceSourcesHealthQuery,
-  useIntelligenceSymbolSetsQuery,
   useIntelligenceUpcomingCatalystsQuery,
   useRunIntelligenceMutation,
   useTestIntelligenceProviderMutation,
   useUpdateIntelligenceConfigMutation,
-  useUpdateIntelligenceSymbolSetMutation,
 } from '@/features/intelligence/hooks';
 import type {
   IntelligenceConfig,
-  IntelligenceLlmProvider,
   IntelligenceRunRequest,
 } from '@/features/intelligence/types';
+import { useIntelligenceSymbolSetEditor } from '@/features/intelligence/useIntelligenceSymbolSetEditor';
+import { PROVIDER_MODELS } from '@/content/intelligenceProviders';
 import { t } from '@/i18n/t';
 
-function normalizeSymbols(input: string): string[] {
-  const seen = new Set<string>();
-  return input
-    .split(',')
-    .map((value) => value.trim().toUpperCase())
-    .filter((value) => {
-      if (!value || seen.has(value)) return false;
-      seen.add(value);
-      return true;
-    });
-}
-
-const PROVIDER_MODELS: Record<IntelligenceLlmProvider, string[]> = {
-  ollama: ['mistral:7b-instruct', 'llama3.1:8b-instruct', 'qwen2.5:7b-instruct'],
-  openai: ['gpt-4o-mini', 'gpt-4.1-mini', 'gpt-4.1', 'o4-mini'],
-  mock: ['mock-classifier'],
-};
-
-const PROVIDER_DEFAULTS: Record<IntelligenceLlmProvider, { model: string; baseUrl: string }> = {
-  ollama: { model: 'mistral:7b-instruct', baseUrl: 'http://localhost:11434' },
-  openai: { model: 'gpt-4o-mini', baseUrl: 'https://api.openai.com/v1' },
-  mock: { model: 'mock-classifier', baseUrl: '' },
-};
 
 export default function IntelligencePage() {
   const navigate = useNavigate();
   const configQuery = useIntelligenceConfigQuery();
   const providersQuery = useIntelligenceProvidersQuery();
-  const symbolSetsQuery = useIntelligenceSymbolSetsQuery();
-
   const updateConfigMutation = useUpdateIntelligenceConfigMutation();
-  const createSymbolSetMutation = useCreateIntelligenceSymbolSetMutation();
-  const updateSymbolSetMutation = useUpdateIntelligenceSymbolSetMutation();
-  const deleteSymbolSetMutation = useDeleteIntelligenceSymbolSetMutation();
   const testProviderMutation = useTestIntelligenceProviderMutation();
 
   const [draftConfig, setDraftConfig] = useState<IntelligenceConfig | null>(null);
   const [jobId, setJobId] = useState<string>();
   const [asofDate, setAsofDate] = useState<string>();
   const [manualSymbolsInput, setManualSymbolsInput] = useState('');
-  const [selectedSymbolSetId, setSelectedSymbolSetId] = useState('');
-  const [symbolSetName, setSymbolSetName] = useState('');
-  const [symbolSetSymbolsInput, setSymbolSetSymbolsInput] = useState('');
-  const [showAdvancedConfig, setShowAdvancedConfig] = useState(() => {
-    if (typeof window === 'undefined') {
-      return false;
-    }
-    return window.localStorage.getItem('intelligence.showAdvancedConfig') === 'true';
-  });
+  const [showAdvancedConfig, setShowAdvancedConfig] = useLocalStorage(
+    'intelligence.showAdvancedConfig',
+    false,
+    // Handle both legacy raw-string ("true") and new JSON (true) formats
+    (val) => val === true || val === 'true',
+  );
+  const {
+    symbolSets,
+    selectedSymbolSetId,
+    setSelectedSymbolSetId,
+    selectedSymbolSet,
+    symbolSetName,
+    setSymbolSetName,
+    symbolSetSymbolsInput,
+    setSymbolSetSymbolsInput,
+    createSymbolSet,
+    updateSymbolSet,
+    deleteSymbolSet,
+    isCreating: isCreatingSymbolSet,
+    isUpdating: isUpdatingSymbolSet,
+  } = useIntelligenceSymbolSetEditor();
 
   useEffect(() => {
     if (configQuery.data) {
@@ -84,13 +67,14 @@ export default function IntelligencePage() {
     }
   }, [configQuery.data]);
 
-  useEffect(() => {
-    window.localStorage.setItem('intelligence.showAdvancedConfig', String(showAdvancedConfig));
-  }, [showAdvancedConfig]);
-
-  const symbolSets = symbolSetsQuery.data?.items ?? [];
-  const selectedSymbolSet = symbolSets.find((item) => item.id === selectedSymbolSetId);
-  const manualSymbols = useMemo(() => normalizeSymbols(manualSymbolsInput), [manualSymbolsInput]);
+  const manualSymbols = useMemo(
+    () => manualSymbolsInput
+      .split(',')
+      .map((v) => v.trim().toUpperCase())
+      .filter(Boolean)
+      .filter((v, i, arr) => arr.indexOf(v) === i),
+    [manualSymbolsInput],
+  );
   const llmModelOptions = useMemo(() => {
     if (!draftConfig) return [];
     const provider = draftConfig.llm.provider;
@@ -100,13 +84,6 @@ export default function IntelligencePage() {
     }
     return options.includes(draftConfig.llm.model) ? options : [draftConfig.llm.model, ...options];
   }, [draftConfig]);
-
-  useEffect(() => {
-    if (selectedSymbolSet) {
-      setSymbolSetName(selectedSymbolSet.name);
-      setSymbolSetSymbolsInput(selectedSymbolSet.symbols.join(', '));
-    }
-  }, [selectedSymbolSet]);
 
   const runMutation = useRunIntelligenceMutation((launch) => {
     setJobId(launch.jobId);
@@ -175,40 +152,6 @@ export default function IntelligencePage() {
     });
   };
 
-  const createSymbolSet = () => {
-    const symbols = normalizeSymbols(symbolSetSymbolsInput);
-    if (!symbolSetName.trim() || symbols.length === 0) return;
-    createSymbolSetMutation.mutate(
-      { name: symbolSetName.trim(), symbols },
-      {
-        onSuccess: (created) => {
-          setSelectedSymbolSetId(created.id);
-        },
-      }
-    );
-  };
-
-  const updateSymbolSet = () => {
-    if (!selectedSymbolSetId) return;
-    const symbols = normalizeSymbols(symbolSetSymbolsInput);
-    if (!symbolSetName.trim() || symbols.length === 0) return;
-    updateSymbolSetMutation.mutate({
-      id: selectedSymbolSetId,
-      payload: { name: symbolSetName.trim(), symbols },
-    });
-  };
-
-  const deleteSymbolSet = (id: string) => {
-    deleteSymbolSetMutation.mutate(id, {
-      onSuccess: () => {
-        if (selectedSymbolSetId === id) {
-          setSelectedSymbolSetId('');
-          setSymbolSetName('');
-          setSymbolSetSymbolsInput('');
-        }
-      },
-    });
-  };
 
   if (configQuery.isLoading || !draftConfig) {
     return <div className="text-sm text-gray-600 dark:text-gray-400">{t('intelligencePage.loading')}</div>;
@@ -295,10 +238,10 @@ export default function IntelligencePage() {
             />
           </div>
           <div className="mt-3 flex flex-wrap gap-2">
-            <Button onClick={createSymbolSetMutation.isPending ? undefined : createSymbolSet}>
+            <Button onClick={isCreatingSymbolSet ? undefined : createSymbolSet}>
               {t('intelligencePage.symbols.createSet')}
             </Button>
-            <Button variant="secondary" onClick={updateSymbolSetMutation.isPending ? undefined : updateSymbolSet}>
+            <Button variant="secondary" onClick={isUpdatingSymbolSet ? undefined : updateSymbolSet}>
               {t('intelligencePage.symbols.updateSet')}
             </Button>
             {selectedSymbolSetId && (
@@ -480,281 +423,16 @@ export default function IntelligencePage() {
       </Card>
 
       {showAdvancedConfig ? (
-        <Card variant="bordered">
-          <CardHeader>
-            <CardTitle>{t('intelligencePage.config.title')}</CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.enabled')}</span>
-                <input
-                  type="checkbox"
-                  checked={draftConfig.enabled}
-                  onChange={(event) =>
-                    setDraftConfig({ ...draftConfig, enabled: event.target.checked })
-                  }
-                  aria-label={t('intelligencePage.config.enabled')}
-                  className="h-5 w-5 rounded border border-gray-300"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.providers')}</span>
-                <input
-                  value={draftConfig.providers.join(', ')}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      providers: event.target.value
-                        .split(',')
-                        .map((value) => value.trim().toLowerCase())
-                        .filter((value, index, list) => value && list.indexOf(value) === index),
-                    })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.llmProvider')}</span>
-                <select
-                  value={draftConfig.llm.provider}
-                  onChange={(event) => {
-                    const nextProvider = event.target.value as IntelligenceLlmProvider;
-                    const defaults = PROVIDER_DEFAULTS[nextProvider];
-                    const nextModel = PROVIDER_MODELS[nextProvider].includes(draftConfig.llm.model)
-                      ? draftConfig.llm.model
-                      : defaults.model;
-                    setDraftConfig({
-                      ...draftConfig,
-                      llm: {
-                        ...draftConfig.llm,
-                        provider: nextProvider,
-                        model: nextModel,
-                        baseUrl: defaults.baseUrl,
-                      },
-                    });
-                  }}
-                  aria-label={t('intelligencePage.config.llmProvider')}
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                >
-                  <option value="ollama">ollama</option>
-                  <option value="openai">openai</option>
-                  <option value="mock">mock</option>
-                </select>
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.llmModel')}</span>
-                <select
-                  value={draftConfig.llm.model}
-                  onChange={(event) =>
-                    setDraftConfig({ ...draftConfig, llm: { ...draftConfig.llm, model: event.target.value } })
-                  }
-                  aria-label={t('intelligencePage.config.llmModel')}
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                >
-                  {llmModelOptions.map((modelName) => (
-                    <option key={modelName} value={modelName}>
-                      {modelName}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.llmBaseUrl')}</span>
-                <input
-                  value={draftConfig.llm.baseUrl}
-                  onChange={(event) =>
-                    setDraftConfig({ ...draftConfig, llm: { ...draftConfig.llm, baseUrl: event.target.value } })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.llmApiKey')}</span>
-                <input
-                  type="password"
-                  autoComplete="off"
-                  value={draftConfig.llm.apiKey}
-                  onChange={(event) =>
-                    setDraftConfig({ ...draftConfig, llm: { ...draftConfig.llm, apiKey: event.target.value } })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-              <label className="text-sm md:col-span-2">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.llmSystemPrompt')}</span>
-                <textarea
-                  value={draftConfig.llm.systemPrompt}
-                  onChange={(event) =>
-                    setDraftConfig({ ...draftConfig, llm: { ...draftConfig.llm, systemPrompt: event.target.value } })
-                  }
-                  rows={4}
-                  placeholder={t('intelligencePage.config.llmSystemPromptPlaceholder')}
-                  className="w-full rounded border border-gray-300 px-3 py-2 font-mono text-xs"
-                />
-              </label>
-              <label className="text-sm md:col-span-2">
-                <span className="mb-1 block text-xs text-gray-500">
-                  {t('intelligencePage.config.llmUserPromptTemplate')}
-                </span>
-                <textarea
-                  value={draftConfig.llm.userPromptTemplate}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      llm: { ...draftConfig.llm, userPromptTemplate: event.target.value },
-                    })
-                  }
-                  rows={8}
-                  placeholder={t('intelligencePage.config.llmUserPromptTemplatePlaceholder')}
-                  className="w-full rounded border border-gray-300 px-3 py-2 font-mono text-xs"
-                />
-                <p className="mt-1 text-xs text-gray-500">{t('intelligencePage.config.llmPromptTemplateHint')}</p>
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.maxConcurrency')}</span>
-                <input
-                  type="number"
-                  min={1}
-                  max={16}
-                  value={draftConfig.llm.maxConcurrency}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      llm: {
-                        ...draftConfig.llm,
-                        maxConcurrency: Math.max(1, Math.min(16, Number(event.target.value) || 1)),
-                      },
-                    })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.lookbackHours')}</span>
-                <input
-                  type="number"
-                  min={1}
-                  value={draftConfig.catalyst.lookbackHours}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      catalyst: {
-                        ...draftConfig.catalyst,
-                        lookbackHours: Math.max(1, Number(event.target.value) || 1),
-                      },
-                    })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">{t('intelligencePage.config.minScore')}</span>
-                <input
-                  type="number"
-                  step={0.01}
-                  min={0}
-                  max={1}
-                  value={draftConfig.opportunity.minOpportunityScore}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      opportunity: {
-                        ...draftConfig.opportunity,
-                        minOpportunityScore: Math.max(0, Math.min(1, Number(event.target.value) || 0)),
-                      },
-                    })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">Enable fallback scraping</span>
-                <input
-                  type="checkbox"
-                  checked={draftConfig.sources.scrapingEnabled}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      sources: {
-                        ...draftConfig.sources,
-                        scrapingEnabled: event.target.checked,
-                      },
-                    })
-                  }
-                  className="h-5 w-5 rounded border border-gray-300"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">Allowed domains (comma-separated)</span>
-                <input
-                  value={draftConfig.sources.allowedDomains.join(', ')}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      sources: {
-                        ...draftConfig.sources,
-                        allowedDomains: event.target.value
-                          .split(',')
-                          .map((value) => value.trim().toLowerCase())
-                          .filter((value, index, list) => value && list.indexOf(value) === index),
-                      },
-                    })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-              <label className="text-sm">
-                <span className="mb-1 block text-xs text-gray-500">Binary event window (days)</span>
-                <input
-                  type="number"
-                  min={1}
-                  max={30}
-                  value={draftConfig.calendar.binaryEventWindowDays}
-                  onChange={(event) =>
-                    setDraftConfig({
-                      ...draftConfig,
-                      calendar: {
-                        ...draftConfig.calendar,
-                        binaryEventWindowDays: Math.max(1, Math.min(30, Number(event.target.value) || 1)),
-                      },
-                    })
-                  }
-                  className="w-full rounded border border-gray-300 px-3 py-2"
-                />
-              </label>
-            </div>
-
-            <div className="mt-4 flex flex-wrap gap-2">
-              <Button onClick={saveConfig} disabled={updateConfigMutation.isPending}>
-                {updateConfigMutation.isPending ? t('intelligencePage.config.saving') : t('intelligencePage.config.save')}
-              </Button>
-              <Button variant="secondary" onClick={testProvider}>
-                {t('intelligencePage.config.testProvider')}
-              </Button>
-            </div>
-
-            {testProviderMutation.data && (
-              <p className="mt-2 text-sm text-gray-600 dark:text-gray-400">
-                {t('intelligencePage.config.providerResult', {
-                  provider: testProviderMutation.data.provider,
-                  status: testProviderMutation.data.available ? 'OK' : 'UNAVAILABLE',
-                })}
-              </p>
-            )}
-
-            {providersQuery.data && (
-              <div className="mt-3 space-y-1 text-xs text-gray-500">
-                {providersQuery.data.map((provider) => (
-                  <p key={provider.provider}>
-                    {provider.provider}: {provider.available ? 'OK' : 'UNAVAILABLE'}
-                    {provider.detail ? ` (${provider.detail})` : ''}
-                  </p>
-                ))}
-              </div>
-            )}
-          </CardContent>
-        </Card>
+        <IntelligenceConfigPanel
+          draftConfig={draftConfig}
+          llmModelOptions={llmModelOptions}
+          isSaving={updateConfigMutation.isPending}
+          testProviderResult={testProviderMutation.data}
+          providersStatus={providersQuery.data}
+          onConfigChange={setDraftConfig}
+          onSave={saveConfig}
+          onTestProvider={testProvider}
+        />
       ) : null}
 
       <div
