+++ web-ui/src/utils/dailyReviewFormatters.ts
@@ -1,0 +1,14 @@
+import { t } from '@/i18n/t';
+
+const TIME_EXIT_REASON_PATTERN = /Time exit:\s*(\d+)\s*bars since entry_date\s*>=\s*(\d+)/i;
+
+export function formatDailyReviewReason(reason: string): string {
+  const timeExitMatch = reason.match(TIME_EXIT_REASON_PATTERN);
+  if (timeExitMatch) {
+    return t('dailyReview.reason.timeExit', {
+      barsSince: Number(timeExitMatch[1]),
+      maxBars: Number(timeExitMatch[2]),
+    });
+  }
+  return reason;
+}
